% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick.R
\name{attachDependenciesCandleStick}
\alias{attachDependenciesCandleStick}
\title{Attach dependencies for candlestick}
\usage{
attachDependenciesCandleStick(tag)
}
\arguments{
\item{tag}{any shiny element}
}
\description{
Adds the dependencies jquery, jquery hammer, hammer, candlestick and font-awesome
   to the candlestick
}
