% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick.R
\name{CandleStick}
\alias{CandleStick}
\title{CandleStick input with three options}
\usage{
CandleStick(id = "fancycolorpicker", left = c(f = "f182"),
  right = c(m = "f183"), default = c(b = "f22d"), fa_mode = TRUE,
  candlestick_mode = "contents", candlestick_size = "md",
  contents_color = "#fc4c02")
}
\arguments{
\item{id}{ID of the element}

\item{left}{named vector with left option - name is the value}

\item{right}{named vector with right option - name value}

\item{default}{named vector with central option}

\item{fa_mode}{\code{logical} Whether to use font-awesome icons, works
just in 'contents' mode}

\item{candlestick_mode}{'option' or 'contents'}

\item{candlestick_size}{Sizes (lg, md (default), sm, xs)}

\item{contents_color}{Color of the select element in case the 'contents' mode is used.}
}
\description{
This is an input taken from https://github.com/EdouardTack/candlestick
to allow a three item candlestick in shiny
}
\references{
\link{registerCandleStick}, \link{attachDependenciesCandleStick},
}
