% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestick.R
\name{updateCandleStick}
\alias{updateCandleStick}
\title{Update a candlestick value}
\usage{
updateCandleStick(inputId, value = "on")
}
\arguments{
\item{inputId}{ID of Canldestick}

\item{value}{on / off / default}
}
\description{
Update a candlestick value
}
